NU-1778 types update
