+++ designer/client/src/components/toolbars/activities/helpers/extendActivitiesWithUIData.test.ts
@@ -2,7 +2,7 @@ import { extendActivitiesWithUIData } from "./extendActivitiesWithUIData";
 import { mergeActivityDataWithMetadata } from "./mergeActivityDataWithMetadata";
 import { sampleMetadataResponse } from "../../../../../__mocks__/fixtures/sampleMetadataResponse";
 import { sampleActivitiesResponse, sampleActivitiesWithRepetiveResponse } from "../../../../../__mocks__/fixtures/sampleActivitiesResponse";
-import { ActivitiesResponse } from "../../../../http/HttpService";
+import { ActivitiesResponse } from "../types";
 
 describe("extendActivitiesWithUIData", () => {
     it("should render activities with metadata when items are not grouped", () => {
