+++ designer/client/src/components/toolbars/activities/useActivitiesSearch.test.ts
@@ -1,7 +1,7 @@
 import { useActivitiesSearch } from "./useActivitiesSearch";
 import { act, renderHook } from "@testing-library/react";
 import { extendActivitiesWithUIData } from "./helpers/extendActivitiesWithUIData";
-import { ActivitiesResponse } from "../../../http/HttpService";
+import { ActivitiesResponse } from "./types";
 import { mergeActivityDataWithMetadata } from "./helpers/mergeActivityDataWithMetadata";
 import { sampleMetadataResponse } from "../../../../__mocks__/fixtures/sampleMetadataResponse";
